+++ app/src/main/java/com/codesoom/assignment/application/UserQueryService.java
@@ -1,0 +1,29 @@
+package com.codesoom.assignment.application;
+
+import com.codesoom.assignment.ResourceNotFoundException;
+import com.codesoom.assignment.domain.User;
+import com.codesoom.assignment.domain.UserRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@Transactional(readOnly = true)
+public class UserQueryService {
+
+    private UserRepository repository;
+
+    public UserQueryService(UserRepository repository) {
+        this.repository = repository;
+    }
+
+    public List<User> findAll(){
+        return repository.findAll();
+    }
+
+    public User findUser(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(id));
+    }
+}
